Russian APT/APT29-Adversary-Simulation/requirements.txt
pyautogui
pyvirtualdisplay
requests
pycryptodome

